mjpeg-streaming/api/requirements.txt
Pillow==9.3.0
opencv-python==4.5.1.48
depthai==2.20.2
blobconverter==1.2.9